+++ cdap-runtime-ext-dataproc/src/main/java/io/cdap/cdap/runtime/spi/provisioner/dataproc/DataprocClientFactory.java
@@ -1,0 +1,36 @@
+/*
+ *  Copyright © 2022 Cask Data, Inc.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ *  use this file except in compliance with the License. You may obtain a copy of
+ *  the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ *  License for the specific language governing permissions and limitations under
+ *  the License.
+ */
+
+package io.cdap.cdap.runtime.spi.provisioner.dataproc;
+
+import io.cdap.cdap.runtime.spi.ssh.SSHPublicKey;
+
+import java.io.IOException;
+import java.security.GeneralSecurityException;
+import javax.annotation.Nullable;
+
+/**
+ * Creates DataprocClients.
+ */
+public interface DataprocClientFactory {
+
+  default DataprocClient create(DataprocConf conf) throws GeneralSecurityException, IOException {
+    return create(conf, false, null);
+  }
+
+  DataprocClient create(DataprocConf conf, boolean requireSSH,
+                        @Nullable SSHPublicKey sshPublicKey) throws IOException, GeneralSecurityException;
+}
